requirements.txt
aiofiles==24.1.0
anyio==4.9.0
audiosegment==0.23.0
blinker==1.9.0
boto3==1.37.35
botocore==1.37.35
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
Deprecated==1.2.18
dotenv==0.9.9
exceptiongroup==1.2.2
Flask==3.1.0
get-annotations==0.1.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-tools==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
numpy==2.0.2
packaging==24.2
protobuf==5.29.4
pycparser==2.22
pydub==0.25.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
typing_extensions==4.13.2
urllib3==1.26.20
webrtcvad==2.0.10
Werkzeug==3.1.3
wrapt==1.17.2
yandex-cloud-ml-sdk==0.8.0
yandex-speechkit==1.5.0
yandexcloud==0.340.0
zipp==3.21.0
